Updated code from my arduino project
also commented the pin connections.

So far, this does work though I'm still having issues with calibration.
I probably need to hook up RV and take this value in to account per
iteration in loop. The temperature pin is also needed but not currently
checked.

For now, I have a simple-ish calibration routine. During this
calibration process, place a cup or something over the sensor as this
thing ( as the doc warns ;) ) very very sensitive!